Create a stable branch for KRoC 1.5.
This branch is for end-users of KRoC to use instead of trunk; it should always
be safe to pull from. Releases will then be tags off this branch (so
/kroc/tags/kroc-1.5.0-pre4 will be a copy of /kroc/branches/kroc-1.5, and the
release tarball will be rolled from the tag).

Prior to the release of KRoC 1.5.0, we should merge changes from trunk to this
branch once we're happy that they're reasonably stable (i.e. after we've tested
them on the build farm, at the bare minimum). After the release, we should only
merge bugfixes from trunk to here.

Developers violating the above rules will be shot.