Right for now I'm going to call my new runtime addition "disconnected channels".  Which makes slightly more sense than "broken", and is less likely to be mistaken for me saying something is actually broken.
The idea of a disconnected channel is like an unconnected wire in a circuit.  It floats, and output to it gets thrown away, and input from it is useless.  This turns on its head the conventional occam semantics for channels, where if no one is listening then we block.  With a disconnected channel we never block.  It is like outputting to /dev/null and reading from /dev/zero.

Understandably breaking semantics which have existed and been tried and tested for >20 years leaves some people nervous.  However, there is for forseeable need for these semantics and a means for algebraic representation for them.  As long as the disconnected channels exist on the boundary between the program and its environment then we can define them as we like.  For the purposes of CSP verification the boundary channels are features of the environment not the program under analysis.  It just happens in our case they are events which always complete when synchronised on by the program.

Anyway, for the moment disconnected channels are an interesting research topic with a practical application between execution contexts in the TVM.  Oh and the code is completely untested ;-)