I've come to the conclusion that the code that assigns workspace to C./B./BX. calls is wrong.  It doesn't provide any storage for the register arguments.  This can actually lead to problems when workspace is tightly packed.
For example take the following occam code:

  PROC no.op()
    SKIP
  :

occ21 will allocate 3 workspace words for this process even though it does nothing.  There words are to store the registers which are saved to the workspace on making a CALL.

Therefore allocating 0 words for a C./B./BX. call is in my opinion an error.  It is of course an error you won't notice if you always do some form of communication or other rescheduling in the same code body as you make an FFI call, as additional space will be available for the Iptr/Link/Pri, etc.