Clean up insert debugging a bit.
Previously, mdparam[0] held the file and line numbers, and mdparam[1] was a
pointer to a function that set mdparam[0] to point to the filename table. This
was a pain because the function needed to have %esi set correctly (and CCSP
didn't do so). Now, mdparam[1] just holds the filename table pointer directly.

I've also split the CCSP insert debugging decoder out into a separate function,
which avoids some duplicated code and makes it easier to use when chasing down
memory leaks (or similar problems).