Implement full sleeping of the Blackfin processor.  At least I believe I've implemented what the Blackfin manuals refer to as sleep: the details are slightly hazy.
The firmware now stops the core clock when idle or waiting for a timeout, in turn it starts up general purpose timer 7 in the peripheral clock domain to act as a high-precision wake up signal.  On waking up, either by the alarm or some other source of activity, the core timer is restore and advanced using the timer 7 counter.

Due to only having 32 bits of precision in the timer 7 counter, we have to wake up every 35 seconds to maintain a valid state.  So even when the timeout is long or the system is completely idle it will wake up briefly every 35 seconds.

Because of bouncing time back and forth between the core timer and GP timer 7, there will probably be some more skew in the clock for heavily idle systems.  Therefore if you require fairly accurate (sub-millisecond) timing you should probably spin, rather than letting the system go idle.

One thing I have noticed is that the JTAG software gets quite upset sometimes when the CPU is stopped, mainly because it wants to stop the processor which is already stopped.  Just trying again usually solves things, but I'll keep an eye on this.  If it proves to be a problem I may need to do something extra.

If I had the correct cables I would now hook up my multimeter and see how much power is being saved by this over just running the CPU flat out...