Start adding support for external channel bundles.  These are mobile type channel bundles for which channel read/write/etc operations are forwarded out some externally defined interface (C functions).
This is very much like the user-defined-channels functionality which KRoC used to have.  It is completely run-time based, so no compiler support is required, operations are effectively transparent to the program.

Underneath this works by using the second lowest bit of the channel word to indicate an extended channel.  The channel contents is a pointer to the mobile type which contains the channel.  The mobile type in turn contains a pointer to the external interface table, and a user defined data pointer.  The run-time detects operations on these external channels, loads the interface table, and makes the appropriate function call.

At present ALTing is not supported.