Rework the code that sets the IA32 FPU control register.
There's now just one function that does both the initial setup and the update
when the rounding mode is changed. Rather than loading and adjusting the
register, it just sets it directly to the new value.

This behaves like the original code in that it always leaves FP exceptions
masked (except for "invalid operation"). I suspect this wasn't the intention
from the "/* let exceptions get generated */" comment that used to be present
-- but since it always ORd the mask into the existing one, it would always have
been at least 0x3e from the initial setup.