- fluke/setup.h Primarily whitespace. In particular, tabs at the beginning of lines are OK, because you can change the tabsize and the code still looks "right". (eg. ts=2 or ts=8 doesn't change code layout). However, tabs in the middle of lines strikes me as bad practice, because when you change the tabstops, the code layout changes. I don't like that at all; I've replaced some tabs that were in the middle of lines with spaces, so the layout isn't effected by changes in ts size.
-fluke/fluke_uart.h
I've moved the UART_BAUD definition to lpc/uart.h, but I'm not happy about it.
I think it's largely because they used a 20MHz crystal whereas the example code
was written against a 14MHz crystal; thus, baud rate calculations are a bit
different. Ultimately, I'd like to see fluke_uart.h and tvm_uart.h merged so
they are clearly specific to this port. I've commented or #ifdef'd out most of
the code that was copy-and-pasted by the Fluke team from RO Software's headers.

-fluke/tools/serial_control.py
I added a 1 second delay. It didn't change anything.

In particular, I've often found that uploading code yields a crashing firmware.
Resetting does nothing---it's a quick reset, and then it keeps crashing.
However, if I pull the power, and bring the board back up, I can then do a
"reset-term" and get firmware that works every time.

There is clearly some buffering in the serial code, but the reset is not doing
the job... 

Trying to figure out why things keep crashing was part of my inclusion of the
"date.h" generation (see Makefile comments). I wanted to be sure of which
firmware I was running, and sure enough, I'm running the most recent. The
firmware I am generating is, I believe, good, but for some reason resetting
(using the wonky lpc21isp) does not actually kick the processor back into a good
state.

Do we need to clear memory? Is there some other kind of setup we're missing? Are
we failing to wait for a crystal to stabilize? Why would we crash constantly
after upload/reset, but not after powercycle/reset?

-fluke/Makefile
I've added a target that generates "date.h". "main.o" depends on it. A "make
clean / make" guarantees that a new "date.h" comes into being, and allows us to
tell for certain that we're not somehow running an old firmware.

-fluke/lpc/config.h
Commented out the FOSC. This is in the Fluke-specific setup, which will be
consolidated someday.

-fluke/lpc/uart.h
Moved UART_BAUD() here.