The RO Software code already had timers set up; we just needed to scale appropriately and call their version of "get_time".
The firmware was just "tick"ing at us, to see if the timer was behaving sanely.

Fixes to the Makefile to get firmeware dependencies right.

Overall, no core changes of substance, but a new piece of functionality is
latched in. Sleep will come later, but that seems straight-forward enough when
the time comes... MERCY! A pun, and really, it wasn't intentional.