Get inmoslibs to build.
This is essentially the same work to occbuild that I just did to kroc:
the OCCAM_OCCBUILD macro can now find occbuild in the source tree, and
operate without needing to run occbuild when doing an in-tree build; and
occbuild has gained a --in-tree option.

inmoslibs itself didn't need any changes. (This is not going to be true
of the other modules, unfortunately, but they should be pretty trivial.)

It's now possible to use this tree to build trivial example programs
(e.g. "hello world" without using the course library).