Make it possible to build modules that depend on other modules.
This is done with some occbuild magic that's activated only for in-tree
builds. When occbuild builds a .module file, it records its location
into $KROC_BUILD_ROOT/in-tree-modules.  The OCCAM_INCLUDE autoconf macro
then looks in that file, and adds --search directives to OCCBUILD for
all the modules it finds. (The directory for forall is built into
the kroc script, since it needs to always be available.)

No changes are necessary to modules other than adding the -I option for
aclocal -- and making sure they check for all the modules they need
using OCCAM_INCLUDE. (You will now find out about this fairly rapidly if
it's not the case...)

This isn't a particularly elegant solution (particularly in terms of the
length of the commands that make ends up executing!), but it's simple.