OK, so the previous inter-module dependency scheme was simple, obvious, and completely wrong. I'd modified occbuild to register modules as they're built -- but that won't work, because we need to be able to find everything at configure time.
Now there's an OCCAM_PROVIDE macro that registers modules in their
configure scripts. This means that packages do need to be modified to
say what they provide, but that's pretty straightforward.

In addition, OCCAM_INCLUDE has been modified to chase module
dependencies. Specifying dependencies is pretty easy because it's
usually anything you've OCCAM_INCLUDEd, but you can override that in
cases where you're providing multiple modules. The code to implement
this in the macro is pretty nasty (recursive shell functions?), but it
seems to work. (Famous last words.)