Fix datadir installation, and clean up some of the muckier Makefiles in the tree in the process.
In particular, demos/game is no longer a crawling horror, and
$(INSTALL_DATA) is no longer used (because dist_whatever_DATA is much
more straightforward and avoids duplication).

In addition, this adds the doc directory to the build, which means it's
now installing everything a regular KRoC installation does for me.
There's still stuff that needs tidying up, but this branch is basically
there now.