Make it possible to build the Transterpreter from this tree.
The top-level configure script now recognises the occbuild
--with-toolchain option, and you can specify either
--with-toolchain=kroc (the default) or --with-toolchain=tvm.
(You can install both toolchains into the same prefix quite happily, but
the build tree can only handle one at a time because the object
filenames will collide.)

Otherwise, this is basically the same thing I did for the kroc script:
skroc has gained a --in-tree option that's used when doing in-tree
builds, and there are autoconf macros to detect Transterpreter build
parameters (both of which are trivial).

This commit fixes an existing bug in skroc: the --tvm_config_h.in
argument would never have worked, because its value was used before the
arguments were parsed. The tvm_config.h-based autoconf is now done
separately after the arguments have been parsed.

The build now fails once it gets into modules that need CIF support, but
that's expected and I'll fix it in a minute. What can be installed so
far appears to work at least as far as being able to compile a "hello,
world" program, although the tvm binary hangs on my machine for reasons
not yet investigated.