Fix a bug with NORM. Fix rounding of float to int to deal with negatives.
One of LLVM's weaknesses is the present lack of support for rounding modes in floating point.
The rounding mode of all operations is however defined to be 'round to zero'.
Therefore to support 'round to nearest' in occam, we generate a (+ 0.5) prior to the round; however, this needs to be (- 0.5) for negative numbers.
Which is what this patch fixes.