Implement LDIV, LMUL, LSHR and LSHL using direct operations on words of the right length in C, when the compiler supports it (which the AVR does).
This adds DWORD/UDWORD defines for signed and unsigned words of the size
that LDIV words upon. All this stuff is hidden behind a TVM_HAVE_DWORD
conditional, since not all N-bit compilers support operations on 2N-bit
words (which is why LDIV was implemented that way in the first place).