Distributed scrolly messages.
This adds some interrupt-driven serial code to ledmatrix, which allows
you to connect multiple ledmatrix instances together over serial links
(i.e. just chaining TX on the first Arduino to RX on the second, and so
on). The result is that you can build a scrolly display as big as you
like, provided you've got enough jumper wires...