Adding tests for CHAN SIGNAL
This commit includes changes to Makefile.am and configure.ac for occ21, as well
as a new "tests" subdirectory. 

We were not sure where tests that cause the compiler to fail should go. They can
always be moved elsewhere, or integrated into another framework as necessary.
For the moment, however, it seemed important to include them. 

A small runner script ("run_occ21_test.py") allows us to write tests that fail
to pass a syntax check, but still "succeed". This was important to us, as we
wanted to make sure that SIGNAL was not able to be used (for example) as part of
a variable declaration; for example,

SIGNAL v:
v := SIGNAL

should fail, because SIGNAL is a channel protocol only, not a value type.

"make check" in occ21 should compile everything and run the tests. Currently, 1
test fails, which involves using SIGNALs under a replicated PAR.