RPM packaging support
There are four files in this commit, and it is unclear how they will grow/change over time. However, they serve as a starting point.

1. tvm.spec / avr.spec
We may have a separate packages for the TVM and AVR builds. For the moment, there are two .spec files, and they will diverge a bit more once I have a better sense for how best to package everything up. I would like to just have one package for the whole project, but that involves more questioning on IRC before I understand how to do multiple configure/make/make install passes in a single spec.

2. vpath-package.py
This is a script that drives the process of checking out the tree, doing a clean build, and generating the RPMs. Likewise, this script will evolve.

3. package-rpm.py
This was the first script I was working with... it is remarkably similar to vpath-package. One of these will go away. For now, I'm committing everything, largely because I am on a VM, and I would like to have the file to go back to if necessary.

I'm aware that there are a number of things that can be cleaned up/improved in the script, but I'm only interested in getting things working at this point, and then I'll worry about pretty.