Exclude libraries for non-POSIX targets
I do not know what the "right" way to do this is. However, this does not appear to break the build, and it isn't a massive change (ie. easily undone manually or rolled back).

I've placed the majority of the libraries under a conditional that requires the POSIX wrapper. Some of these libraries (course? useful?) could be pulled out... but not at this exact moment. I suspect there is a better way to avoid "conditional hell" in the configure.ac, but I don't know what it is.

For now, this gets the AVR building cleanly without breaking the TVM and KRoC (as far as I can tell).