Add functions to generate REAL32s with uniform and Gaussian distributions.
The Gaussian distribution code uses a Box-Muller transform to generate
two Gaussian numbers from two (or more) uniform numbers. The Numerical
Recipes GASDEV which Teodor based his code on does the same thing, but
is more complicated because it's tracking which of the two numbers has
already been used. There are better ways to do this (e.g. the Ziggurat
algorithm) that might be worth considering if it turns out to be a
bottleneck in real simulations.