Export _occ_SDL_make_surface for Windows DLLs.
This is a right pain, because in theory you have to do this explicitly
for *every* symbol you want visible in a DLL. It hadn't been a problem
for the Transterpreter in the past because SWIG does this automatically
for all its wrappers. Perhaps this'd be a good excuse to add an
occam_ffi.h somewhere...

Anyway, SDL applications now work again in a MinGW tvm.exe (at least,
cross-built from Linux and run in WINE; I haven't tried on a real
Windows machine).