Do analog to digital conversions in parallel
This runs ADC on all six "analog in" pins on the Arduino. Unlike the
interrupt-driven ADC, this does not crash anything. Further, it seems to
work---I only have two pots to test with at the moment, but both read on their
respective pins sanely. The rest float unpredictably, which I would expect.