A silly little compilation driver.
I don't really want this in the tree, but it is the little script I'm 
using to support my students on Linux at the moment. There are too many 
parameters, etc., for compiling and uploading programs to the Arduino 
for easy use under Linux.

Assuming that everything is in your path (a big assumption, which I've 
taken care of in our labs), then this "just works".

This is not a permanent fix, and I want this to go away. Or, it should 
be done right. That is, however, down the road. For now, I need to not 
loose this/have it available.