Adding TWI module and test code
This represents my first attempt at implementing a protocol in occam-pi. Comments welcome.

So far, I think I'm talking to my external EEPROM... that said, my session goes like this:

1. Write data (all with OK statuses)
2. Read data (all with OK status)
3. Data always comes back as 43, regardless of what was written.

So, I'm not yet sure that things are working. That said, if I power off the EEPROM, and run my code, it does fail... so the Arduino is (I think) talking to the EEPROM. Perhaps.

More reading and debugging is necessary.