TWI implementation, continued
I'm looking at the documentation for TWI on the Arduino as well as for the 24LC256 EEPROM. I believe, but am not sure, that I'm on the right track.

It looks like I'm successfully writing (at least, I'm not getting any error codes), but I'm falling down on the read. Currently, a random read from the EEPROM crashes out when I send the second START condition... despite the fact that the documentation says I should do just that.

I shall continue...