Storing and retrieving data.
This isn't pretty. 

First, I wonder if the protocol should have ACK and NACK explicitly in the channel protocol. These are just bit operations on the TWI registers... but there are cases where you want to ACK, and cases where you want to NACK. Currently, I've added it as a field in the 'read' protocol (and probably need it elsewhere... I'm implicitly ACKing everything else). As I've implemented the code, I've come to see how the state machine works (it isn't complex), and I've started to understand something about the way client/server protocols are written in o-pi. 

Second... I don't like the fact that I seem (need to re-read) to be talking to the EEPROM in a manner that is not consistent with the part's documentation. That said, it is storing the data I send it and giving it back to me. This may be because it is a very forgiving device. Or, perhaps I have it right. Between the ATmega328 documentation and the 24LC256 documentation, it feels like there are... points where they aren't saying the same thing.

For now, though, this commit has the start of a working TWI protocol. It isn't complete (no support for the ATmega being a slave, for example), and at this point it might be useful to have another device to test on. I'm not sure that I have any other I2C devices laying around, though...