Looking for comments.
1. Slowed TVM output down to 4800. Adjust 'read-arduino' accordingly.

2. Added "softserial.module," which wraps a setup PROC as well as a proc to transmit using the softserial_write FFI.

3. Added an FFI call (slammed in for the moment in an ugly place -- ffi.c) that, in theory, does a serial transmit on a pin of the user's choosing (pins 2 - 7)

I also grossly bumped the location of usercode to 0x5200 for testing purposes. Specifically, the firmware grew a bit.

Using testtx.occ, I transmit values over the software serial link. I connect pin 3 on the transmitter to the RX pin on another Arduino. The receiver is running testrx.occ, which is using the interrupt-based serial routines that were in ledmatrix.occ. 

Running at 57600, I can use the interrupt-based routines (both tx and rx) to talk to each-other. This is good. I can do the same with those slowed down. Also good. I cannot do this, though, when I use my software bitbang code.

Any thoughts? Basically, I get pretty bogus values at the receiving end. They do increase linearly, but they're... wrong. Basically, the bottom half of the nibble seems to be correct, and the upper half of the nibble seems to be wrong.

A bit stuck, really. We're about to ramp up on Mega support (which will give us more serial ports), but in the meantime I need a way for the students to send commands to a GPS while reading from it on the RX pin, and sending debug data back to the PC on the TX pin. Other brainstorms as to how solve that problem are welcome as well...