More bug fixes. This one has a lot to do with what's in wiring.module, and how a lot of it feels like stuff that should not be in wiring.module...
Specifically, pinMode (which makes calls to ports[..]) is declaired in wiring.module, but I was trying to use it in pwm.module.
This was something that shouldn't have worked, as pwm.module doesn't #INCLUDE anything (being one of them lower-down modules). I'm mostly comfortable letting people (read: myself) run into forgetting to set their PWM pins to the output mode, but it would be nice to automate that...

What's important though, is that it's my understanding that wiring and plumbing -- being higher-level modules -- shouldn't make calls to ports[..]. pinMode, digitalRead and digitalWrite all make calls to ports[..]. Problem? Yes? No?