Added constant arrays to m328/constants.module so that moving usart.module to a common directory (for now ../include. Should be moved to something better-named soon) makes serial `just work'.
The reson for doing this is to capitolize on the commonality in the usart code -- attempting to make the whole library more of an API which gets board-specific at the lowest levels possible (in this case, in constants... kindof). I'm honestly not sure how well this will work if we move to support more architectures but, for now, it gets rid of a great deal of repeaded code.
Going to move to pull pwm apart soon.