Added a boardmap.module to each platform to provide an easy board-pin to processor-pin mapping function. Also renamed the dir seeedmega33/ to seeduinomega/, because (if I'm read the commit logs right), we're not using it to develop 3.3v.. things?
This is the first big step in the latest set of changes geared towards making the structure of these modules more intuitive and in-line with the platform-independence we're hoping to achieve with the higher-level modules (plumbing, most significantly). From here we're going to re-write the architecure-specific code so their PROCs take processor pins -- instead of board pins -- as arguments.