This merges the avr-mega branch back into trunk.
For the past few weeks, we've been pushing pretty hard on the Atmel ATmega line of processors... specifically, the ATmega328 and the ATmega1280, which are used on the Arduino and Arduino Mega.

There are still some breakages in the occam code for the Arduino, but we want to get things integrated back into trunk for purposes of packaging. Further, those code breakages will not interfere with packaging, just with the end-user's ability to use the Arduino. (Put another way, we just broke trunk with respect to the Arduino. Sorry.)

The firmware is still good, but the Plumbing support library is temporarily out-of-service. This is only because I needed to merge, and Drew was in the middle of working on serial. Given that we're working on this most intensely right now, this break shouldn't be an issue.

OK. Time to commit, and then I'll begin moving some things around before turning people loose on the tree again.