When a user fires up occPlug for the first time, we insert the correct MODE line in the catalog if it isn't there already.
If a file (*.occ or *.module) was open previously, and jEdit thinks it is in "text" mode, you can "Reload" or "Reload All" under the File menu, and it will refresh the syntax highlighting bindings to properly reflect the file type.

This should be tested under Windows/Mac. It should do nothing there.