Apply changes from ats, fixing the way the TailCall plugin is built.
From ats:

Fix how the shared version of libLLVMTailCallFix is built: list the
source file in SOURCES, not LIBADD.  This means it passes the right
flags to the compiler (e.g. finding the LLVM headers), and it'll work in
a separate object directory.

Caveat: the flags required to build an object in a static library may be
different from those for an object in a shared library (the latter
requires -fPIC on some Linux architectures, for example).