Adding a "bare644" platform.
I'm surprised we didn't have something like this for the 328. Although silly in a way,
this keeps our "board" vs. "architecture" division clean. A "bare644" is a 644p on a 
breadboard, or any configuration where the layout on the board is identical to the layout
on the chip itself. Hence, most of the "mapping" taking place is 1:1.